Create first deployment: 2 apps in 3 app servers
In this version, the docker-compose topology consists of 3 services:
- glassfish
- tomcat
- wildfly

The code of 2 simple apps is provided. The first one contains only
static content (no Java code). The second one contains a very simple
implementation of the MVC patter (with a servlet, a filter and a JSP).

The .war files have been copied in the apps-dist directory, so that they
are available even if the project is not re-built.